Add files via upload
Task 1 
Consider the following moving average trading strategy: Let 7MA be the 7-days moving average, and let 14MA be the 14-days moving average. If the 7MA crosses the 14MA from below, then buy your entire budget. If the 7MA crosses the 14MA from above, then sell your entire portfolio. Your task is to implement the above strategy and test it with MATLAB on the JustEat6M stock’s daily closing prices . Assume you have £1M available to invest and that only integer quantities of the asset can be traded.

Task 2
Consider the following dataset with 8 datapoints and three attributes/features. The target values are captured in the decision column.
	Stock price	Market	News	Decision
1	Up	Down	Impartial	Buy
2	Up	Up	Negative	Buy
3	Up	Stable	Positive	Buy
4	Stable	Up	Negative	Buy
5	Stable	Stable	Negative	Buy
6	Down	Stable	Positive	Buy
7	Down	Down	Positive	Sell
8	Down	Stable	Negative	Sell

Generate a decision tree that corresponds to the above dataset using the fitctree command in MATLAB. 
Consider the fitctree command with the ‘MinParentSize’ argument. Write MATLAB code that computes the maximum value for the ‘MinParentSize’ argument that would yield zero classification loss, and outputs the corresponding value and the tree in graph mode .